+++ packages/mobx-state-tree/src/core/type/index.ts
@@ -1,3 +1,2 @@
 export * from "./type"
-export * from "./type-flags"
 export * from "./type-checker"
