+++ packages/mobx-state-tree/src/core/type/type-checker.ts
@@ -116,4 +116,4 @@ export function typecheckPublic(type: IType<any, any>, value: any): void {
 import { IType } from "./type"
 import { fail, EMPTY_ARRAY, isPrimitive } from "../../utils"
 import { getStateTreeNode, isStateTreeNode } from "../node/node-utils"
-import { isPrimitiveType } from "./type-flags"
+import { isPrimitiveType } from "../../types/primitives"
