+++ packages/mobx-state-tree/src/core/type/type.ts
@@ -1,5 +1,24 @@
 import { action } from "mobx"
-import { TypeFlags } from "./type-flags"
+
+export enum TypeFlags {
+    String = 1 << 0,
+    Number = 1 << 1,
+    Boolean = 1 << 2,
+    Date = 1 << 3,
+    Literal = 1 << 4,
+    Array = 1 << 5,
+    Map = 1 << 6,
+    Object = 1 << 7,
+    Frozen = 1 << 8,
+    Optional = 1 << 9,
+    Reference = 1 << 10,
+    Identifier = 1 << 11,
+    Late = 1 << 12,
+    Refinement = 1 << 13,
+    Union = 1 << 14,
+    Null = 1 << 15,
+    Undefined = 1 << 16
+}
 
 export interface ISnapshottable<S> {}
 
@@ -195,6 +214,10 @@ export abstract class Type<S, T> extends ComplexType<S, T> implements IType<S, T
     removeChild(node: INode, subpath: string): void {
         return fail(`No child '${subpath}' available in type: ${this.name}`)
     }
+}
+
+export function isType(value: any): value is IType<any, any> {
+    return typeof value === "object" && value && value.isType === true
 }
 
 import { EMPTY_ARRAY, fail, isMutable } from "../../utils"
