+++ packages/mobx-state-tree/src/types/utility-types/frozen.ts
@@ -8,7 +8,8 @@ import {
     IValidationResult,
     typeCheckSuccess,
     typeCheckFailure,
-    TypeFlags
+    TypeFlags,
+    isType
 } from "../../core"
 
 export class Frozen<T> extends Type<T, T> {
@@ -65,3 +66,7 @@ export class Frozen<T> extends Type<T, T> {
  * @alias types.frozen
  */
 export const frozen: ISimpleType<any> = new Frozen()
+
+export function isFrozenType(type: any): type is Frozen<any> {
+    return isType(type) && (type.flags & TypeFlags.Frozen) > 0
+}
